+++ changelog.md
@@ -5,6 +5,7 @@
 - Add `ScheduledEvent.image_url`.
 - Add `ScheduledEvent.image_url_as`.
 - Remove extra redirect in interaction client methods, since it could(?) confuse flow order.
+- Add `run_console_till_interruption` (experimental).
 
 #### Bug fixes